requirements.txt
django
Pillow
requests
google-api-core
google-auth
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-cloud-translate
google-cloud-vision
google-resumable-media
googleapis-common-protos
grpcio